rename app and library
closes #17
