+++ angular.json
@@ -3,7 +3,7 @@
   "version": 1,
   "newProjectRoot": "projects",
   "projects": {
-    "ngx-d3timeline": {
+    "ngx-d3timeline-demo": {
       "projectType": "application",
       "schematics": {
         "@schematics/angular:component": {
@@ -17,7 +17,7 @@
         "build": {
           "builder": "@angular-devkit/build-angular:browser",
           "options": {
-            "outputPath": "dist/ngx-d3timeline",
+            "outputPath": "dist/ngx-d3timeline-demo",
             "index": "src/index.html",
             "main": "src/main.ts",
             "polyfills": "src/polyfills.ts",
@@ -66,18 +66,18 @@
         "serve": {
           "builder": "@angular-devkit/build-angular:dev-server",
           "options": {
-            "browserTarget": "ngx-d3timeline:build"
+            "browserTarget": "ngx-d3timeline-demo:build"
           },
           "configurations": {
             "production": {
-              "browserTarget": "ngx-d3timeline:build:production"
+              "browserTarget": "ngx-d3timeline-demo:build:production"
             }
           }
         },
         "extract-i18n": {
           "builder": "@angular-devkit/build-angular:extract-i18n",
           "options": {
-            "browserTarget": "ngx-d3timeline:build"
+            "browserTarget": "ngx-d3timeline-demo:build"
           }
         },
         "test": {
@@ -114,48 +114,48 @@
           "builder": "@angular-devkit/build-angular:protractor",
           "options": {
             "protractorConfig": "e2e/protractor.conf.js",
-            "devServerTarget": "ngx-d3timeline:serve"
+            "devServerTarget": "ngx-d3timeline-demo:serve"
           },
           "configurations": {
             "production": {
-              "devServerTarget": "ngx-d3timeline:serve:production"
+              "devServerTarget": "ngx-d3timeline-demo:serve:production"
             }
           }
         }
       }
     },
-    "@caci/ngx-d3timeline": {
+    "ngx-d3timeline": {
       "projectType": "library",
-      "root": "projects/caci/ngx-d3timeline",
-      "sourceRoot": "projects/caci/ngx-d3timeline/src",
+      "root": "projects/ngx-d3timeline",
+      "sourceRoot": "projects/ngx-d3timeline/src",
       "prefix": "lib",
       "architect": {
         "build": {
           "builder": "@angular-devkit/build-ng-packagr:build",
           "options": {
-            "tsConfig": "projects/caci/ngx-d3timeline/tsconfig.lib.json",
-            "project": "projects/caci/ngx-d3timeline/ng-package.json"
+            "tsConfig": "projects/ngx-d3timeline/tsconfig.lib.json",
+            "project": "projects/ngx-d3timeline/ng-package.json"
           },
           "configurations": {
             "production": {
-              "tsConfig": "projects/caci/ngx-d3timeline/tsconfig.lib.prod.json"
+              "tsConfig": "projects/ngx-d3timeline/tsconfig.lib.prod.json"
             }
           }
         },
         "test": {
           "builder": "@angular-devkit/build-angular:karma",
           "options": {
-            "main": "projects/caci/ngx-d3timeline/src/test.ts",
-            "tsConfig": "projects/caci/ngx-d3timeline/tsconfig.spec.json",
-            "karmaConfig": "projects/caci/ngx-d3timeline/karma.conf.js"
+            "main": "projects/ngx-d3timeline/src/test.ts",
+            "tsConfig": "projects/ngx-d3timeline/tsconfig.spec.json",
+            "karmaConfig": "projects/ngx-d3timeline/karma.conf.js"
           }
         },
         "lint": {
           "builder": "@angular-devkit/build-angular:tslint",
           "options": {
             "tsConfig": [
-              "projects/caci/ngx-d3timeline/tsconfig.lib.json",
-              "projects/caci/ngx-d3timeline/tsconfig.spec.json"
+              "projects/ngx-d3timeline/tsconfig.lib.json",
+              "projects/ngx-d3timeline/tsconfig.spec.json"
             ],
             "exclude": [
               "**/node_modules/**"
@@ -165,7 +165,7 @@
       }
     }
   },
-  "defaultProject": "ngx-d3timeline",
+  "defaultProject": "ngx-d3timeline-demo",
   "cli": {
     "analytics": "8ed8209e-4051-478e-a468-a88c564caca4"
   }
