+++ package.json
@@ -1,10 +1,10 @@
 {
-  "name": "ngx-d3timeline",
+  "name": "ngx-d3timeline-demo",
   "version": "0.0.0",
   "scripts": {
     "ng": "ng",
     "start": "ng serve",
-    "build:lib": "ng build @caci/ngx-d3timeline",
+    "build:lib": "ng build ngx-d3timeline",
     "build": "ng build",
     "test": "jest",
     "lint": "ng lint",
@@ -51,7 +51,7 @@
     "preset": "jest-preset-angular",
     "roots": [
       "src",
-      "projects/caci/ngx-d3timeline"
+      "projects/ngx-d3timeline"
     ],
     "transform": {
       "^.+\\.(ts|js|html)$": "ts-jest"
