+++ projects/ngx-d3timeline/ng-package.json
@@ -1,6 +1,6 @@
 {
   "$schema": "../../../node_modules/ng-packagr/ng-package.schema.json",
-  "dest": "../../../dist/caci/ngx-d3timeline",
+  "dest": "../../../dist/ngx-d3timeline",
   "lib": {
     "entryFile": "src/public-api.ts"
   },
