+++ projects/ngx-d3timeline/src/lib/ngx-d3timeline.component.spec.ts
