+++ projects/ngx-d3timeline/src/public-api.ts
