+++ projects/ngx-d3timeline/tsconfig.lib.json
