+++ projects/ngx-d3timeline/tsconfig.lib.prod.json
