+++ projects/ngx-d3timeline/tsconfig.spec.json
