+++ projects/ngx-d3timeline/tslint.json
