+++ src/app/app.module.ts
@@ -2,7 +2,7 @@ import { BrowserModule } from '@angular/platform-browser';
 import { NgModule } from '@angular/core';
 
 import { AppComponent } from './app.component';
-import { NgxD3timelineModule } from '@caci/ngx-d3timeline';
+import { NgxD3timelineModule } from 'ngx-d3timeline';
 
 @NgModule({
   declarations: [AppComponent],
