+++ tsconfig.json
@@ -11,17 +11,12 @@
     "moduleResolution": "node",
     "importHelpers": true,
     "target": "es2015",
-    "typeRoots": [
-      "node_modules/@types"
-    ],
-    "lib": [
-      "es2018",
-      "dom"
-    ],
+    "typeRoots": ["node_modules/@types"],
+    "lib": ["es2018", "dom"],
     "paths": {
-      "@caci/ngx-d3timeline": [
-        "dist/caci/ngx-d3timeline/caci-ngx-d3timeline",
-        "dist/caci/ngx-d3timeline"
+      "ngx-d3timeline": [
+        "dist/ngx-d3timeline/ngx-d3timeline",
+        "dist/ngx-d3timeline"
       ]
     }
   },
